refactor: move TypeScript types to "index.d.ts" file
This allows for simpler package.json syntax
