+++ lib/index.d.ts
